refactor: restructure example
