+++ examples/ec-site/src/components/ImageCarousel.vue
