+++ examples/ec-site/src/components/Layout.vue
