+++ examples/ec-site/src/pages/detail.vue
