+++ examples/ec-site/src/pages/index.vue
