+++ examples/ec-site/src/prod.ts
@@ -1,7 +1,7 @@
 import { serve } from '@hono/node-server'
 import { serveStatic } from '@hono/node-server/serve-static'
 
-import { app } from './server.ts'
+import { app } from './app/server.ts'
 
 app.use('/assets/*', serveStatic({ root: 'dist/client' }))
